modifed merge submission to avoid omitting fields moved merge-submission and submission-to-clinical to a folder added auth_disabled for local dev flow
