+++ .env.example
@@ -11,6 +11,8 @@ CLINICAL_DB_PASSWORD=
 ############
 # Jwt      #
 ############
+# for local testing only if skipping auth is desired.
+AUTH_DISABLED=false
 JWT_TOKEN_PUBLIC_KEY_URL=
 # optional, if this is set, it will take precedence over fetching from the url
 # if you provide the url instead this env key will be set in the process env from the url
