+++ test/integration/clinical/clinical.spec.ts
@@ -53,6 +53,9 @@ describe('clinical Api', () => {
           mongoPassword() {
             return '';
           },
+          authDisabled() {
+            return false;
+          },
           mongoUser() {
             return '';
           },
